// services/api/src/core/mongo/mongo.enums.ts
export enum MongoErrors {
    Conflict = 'conflict',
    NotFound = 'not-found',
    InvalidObjectId = 'invalid-object-id',
}

export enum ModelIdTag {
    ApiRequest = 'req',
    Balance = 'bal',
    BalanceHistory = 'bal.h',
    Business = 'bz',
    User = 'usr',
    Customer = 'cus',
    Webhook = 'wh',
    Account = 'ac',
    IsvAccount = 'isv.ac',
    AccountLog = 'ac.log',
    CardProgram = 'c.prg',
    CardBin = 'c.bin',
    CardBinAccountLog = 'bin.ac.log',
    Card = 'c',
    CardTransaction = 'c.txn',
    CardAuthorization = 'c.auth',
    Dispute = 'd',
    Payment = 'p',
    PaymentRequest = 'p.req',
    PaymentAuthorization = 'p.auth',
    Event = 'evt',
    EventAttempt = 'evt.a',
    Request = 'req',
    Log = 'log',
    Merchant = 'mch',
    Configuration = 'cfg',
    Fee = 'fee',
    BillProduct = 'bp',
    Billing = 'bil',
    Invoice = 'inv',
    SecretKey = 'sk',
}